feat: Refactor social sign on. Use social sign on within connected accounts area
