+++ blocks/identity-block/components/SocialSignOn/_children/GoogleSignIn.jsx
@@ -1,0 +1,9 @@
+import React from 'react';
+
+function GoogleSignIn() {
+  return (
+    <div id="google-sign-in-button" />
+  );
+}
+
+export default GoogleSignIn;
