+++ blocks/identity-block/components/SocialSignOn/default.test.jsx
@@ -1,6 +1,5 @@
 import React from 'react';
 import { mount, shallow } from 'enzyme';
-import { act } from 'react-dom/test-utils';
 import SocialSignOn from './index';
 import useIdentity from '../Identity';
 
@@ -25,6 +24,9 @@ describe('Subscriptions Social Login Component', () => {
           googleClientId: false,
           facebookAppId: false,
         },
+        initGoogleLogin: () => {},
+        initFacebookLogin: () => {},
+        initializeFacebook: () => {},
       },
     }));
 
@@ -43,10 +45,13 @@ describe('Subscriptions Social Login Component', () => {
           googleClientId: true,
           facebookAppId: false,
         },
+        initGoogleLogin: () => {},
+        initFacebookLogin: () => {},
+        initializeFacebook: () => {},
       },
     }));
 
-    const wrapper = shallow(
+    const wrapper = mount(
       <SocialSignOn onError={() => null} redirectURL="#" />,
     );
     expect(wrapper.find('#google-sign-in-button')).toHaveLength(1);
@@ -62,10 +67,13 @@ describe('Subscriptions Social Login Component', () => {
           googleClientId: false,
           facebookAppId: true,
         },
+        initGoogleLogin: () => {},
+        initFacebookLogin: () => {},
+        initializeFacebook: () => {},
       },
     }));
 
-    const wrapper = shallow(
+    const wrapper = mount(
       <SocialSignOn onError={() => null} redirectURL="#" />,
     );
     expect(wrapper.find('#google-sign-in-button')).toHaveLength(0);
@@ -82,12 +90,9 @@ describe('Subscriptions Social Login Component', () => {
           signinRecaptcha: false,
           recaptchaSiteKey: '6LdXKVQcAAAAAO2tv3GdUbSK-1vcgujX6cP0IgF_',
         })),
-        initGoogleLogin: jest.fn(async () => Promise.resolve({
-          init: true,
-        })),
-        initializeFacebook: jest.fn(async () => Promise.resolve({
-          init: true,
-        })),
+        initGoogleLogin: () => {},
+        initFacebookLogin: () => {},
+        initializeFacebook: () => {},
         configOptions: {
           googleClientId: true,
           facebookAppId: true,
@@ -95,41 +100,9 @@ describe('Subscriptions Social Login Component', () => {
       },
     }));
 
-    const wrapper = shallow(
+    const wrapper = mount(
       <SocialSignOn onError={() => null} redirectURL="#" />,
     );
     expect(wrapper.html()).not.toBe(null);
-  });
-
-  it('renders placeholders for Google and Facebook sign-in buttons', () => {
-    let wrapper;
-    useIdentity.mockImplementation(() => ({
-      isInitialized: true,
-      isLoggedIn: () => true,
-      Identity: {
-        isLoggedIn: jest.fn(async () => false),
-        getConfig: jest.fn(async () => Promise.resolve({
-          signinRecaptcha: false,
-          recaptchaSiteKey: '6LdXKVQcAAAAAO2tv3GdUbSK-1vcgujX6cP0IgF_',
-        })),
-        initGoogleLogin: jest.fn(async () => Promise.resolve({
-          init: true,
-        })),
-        initializeFacebook: jest.fn(async () => Promise.resolve({
-          init: true,
-        })),
-        configOptions: {
-          googleClientId: true,
-          facebookAppId: true,
-        },
-      },
-    }));
-    act(() => {
-      wrapper = mount(
-        <SocialSignOn onError={() => null} redirectURL="#" />,
-      );
-      expect(wrapper.find('#google-sign-in-button')).toHaveLength(1);
-      expect(wrapper.find('.fb-login-button')).toHaveLength(1);
-    });
   });
 });
