+++ blocks/identity-block/features/account-management/_children/social-editable-field-styles.scss
@@ -12,14 +12,3 @@
   line-height: 1.5;
   font-size: calculateRem(16px);
 }
-
-.social-field--container {
-  padding: 1rem;
-  text-align: center;
-  display: flex;
-  justify-content: center;
-}
-
-.social-field--label-text-container {
-  padding-left: calculateRem(8px);
-}
